have all the fields passing Ineed to make fertilizer calulations. Math is commented out in the helper component for dry forulations.
